+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: (LI.FI) Add SUI blockchain support for swaps via Aftermath and Bluefin7k DEXs
+- changed: Include `orderUri` on Bridgeless swaps
 - fixed: Fix `bridgeless` Zano burn asset params
 
 ## 2.33.0 (2025-09-09)